Revert "PRO-2833: Introduce 5 second timeout in external api calls"
