+++ build.gradle
@@ -209,10 +209,6 @@ dependencies {
   testContractCompile group: 'io.jsonwebtoken', name: 'jjwt', version: '0.9.0'
   testContractCompile sourceSets.main.runtimeClasspath
   testContractCompile sourceSets.test.runtimeClasspath
-
-  testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test'
-  testCompile "com.github.tomakehurst:wiremock-standalone:2.17.0"
-
 }
 
 compileJava {
