+++ src/main/java/uk/gov/hmcts/probate/SubmitApplication.java
@@ -14,4 +14,14 @@ public class SubmitApplication {
         SpringApplication.run(SubmitApplication.class, args);
     }
 
+    @Bean
+    @ConfigurationProperties(prefix = "mail")
+    public JavaMailSenderImpl javaMailSender() {
+        return new JavaMailSenderImpl();
+    }
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
 }
